feat: implementando a biblioteca sweet alert

+++ webapp/assets/js/publicacoes.js
@@ -33,9 +33,23 @@ function criarPublicacao(event) {
             conteudo,
         }
     }).done(function() {
-        window.location = "/home";
+        Swal.fire({
+            'title': 'Sucesso!',
+            'text': 'Publicação criada com sucesso!',
+            'icon': 'success',
+            'timer': 2000,
+        }).then((result) => {
+            if (result.isConfirmed) {
+                window.location = "/home";
+            }
+        })
     }).fail(function() {
-        alert("Erro ao criar publicação!");
+        Swal.fire({
+            'title': 'Erro!',
+            'text': 'Erro ao criar publicação!',
+            'icon': 'error',
+            'timer': 2000,
+        })
     });
 }
 
@@ -59,7 +73,12 @@ function curtirPublicacao(event) {
         elementoClicado.addClass('text-danger');
         elementoClicado.removeClass('curtir-publicacao');
     }).fail(function() {
-        alert("Erro ao curtir a publicação");
+        Swal.fire({
+            'title': 'Erro!',
+            'text': 'Erro ao curtir publicação!',
+            'icon': 'error',
+            'timer': 2000,
+        })
     }).always(function() {
         elementoClicado.prop('disabled', false);
     });
@@ -86,7 +105,12 @@ function descurtirPublicacao(event) {
         elementoClicado.addClass('curtir-publicacao');
 
     }).fail(function() {
-        alert("Erro ao descurtir a publicação");
+        Swal.fire({
+            'title': 'Erro!',
+            'text': 'Erro ao descurtir publicação!',
+            'icon': 'error',
+            'timer': 2000,
+        })
     }).always(function() {
         elementoClicado.prop('disabled', false);
     });
@@ -110,9 +134,23 @@ function atualizarPublicacao(event) {
             conteudo: publicacaoConteudo
         }
     }).done(function() {
-        window.location = "/home";
+        Swal.fire({
+            'title': 'Sucesso!',
+            'text': 'Publicação atualizada com sucesso!',
+            'icon': 'success',
+            'confirmButtonText': 'Ok',
+            'confirmButtonColor': '#0d6efd'
+        }).then((result) => {
+            window.location = "/home";
+        })
     }).fail(function() {
-        alert("Erro ao atualizar publicação");
+        Swal.fire({
+            'title': 'Erro!',
+            'text': 'Erro ao atualizar publicação',
+            'icon': 'error',
+            'confirmButtonText': 'Ok',
+            'confirmButtonColor': '#dc3545'
+        })
     }).always(function() {
         elementoClicado.prop('disabled', false);
     });
@@ -121,22 +159,41 @@ function atualizarPublicacao(event) {
 function deletarPublicacao(event) {
     event.preventDefault();
 
-    const elementoClicado = $(event.target);
-    const publicacao = elementoClicado.closest('div');
-    const publicacaoId = publicacao.data('publicacao-id');
+    Swal.fire({
+        'title': 'Atenção!',
+        'text': 'Tem certeza que deseja deletar esta publicação?',
+        'icon': 'warning',
+        'showCancelButton': true,
+        'confirmButtonColor': '#0d6efd',
+        'cancelButtonColor': '#dc3545',
+        'confirmButtonText': 'Sim',
+        'cancelButtonText': 'Cancelar'
+    }).then((result) => {
+        if (!result.isConfirmed) return;
 
-    elementoClicado.prop('disabled', true);
+        const elementoClicado = $(event.target);
+        const publicacao = elementoClicado.closest('div');
+        const publicacaoId = publicacao.data('publicacao-id');
 
-    $.ajax({
-        url: `/publicacoes/${publicacaoId}`,
-        method: "DELETE"
-    }).done(function() {
-        publicacao.fadeOut("slow", function() {
-            $(this).remove();
+        elementoClicado.prop('disabled', true);
+
+        $.ajax({
+            url: `/publicacoes/${publicacaoId}`,
+            method: "DELETE"
+        }).done(function() {
+            publicacao.fadeOut("slow", function() {
+                $(this).remove();
+            });
+        }).fail(function() {
+            Swal.fire({
+                'title': 'Erro!',
+                'text': 'Erro ao deletar publicação',
+                'icon': 'error',
+                'confirmButtonText': 'Ok',
+                'confirmButtonColor': '#dc3545'
+            })
+        }).always(function() {
+            elementoClicado.prop('disabled', false);
         });
-    }).fail(function() {
-        alert("Erro ao deletar publicação");
-    }).always(function() {
-        elementoClicado.prop('disabled', false);
-    });
+    })
 }
